Update cloned version of WhosOnline
The REL1_39 version has broken SQL, the fix for which (T328852) was only added
to the master branch at the time. While it is available on the REL1_40 branch
now, I believe it is preferable to use the master branch rather than a
release-specific branch that is different from the targeted release (1.39).

Clone the (current) latest version of the master branch to include subsequent
localization updates and other improvements since we are upgrading anyway.