Update to PHP 8.1 (#181)
Switches the base image to `debian:12.5` rather than `debian:11.7`. The PHP
packages are retrieved from packages.sury.org - all of the existing php7.4
packages are switched to php8.1, and additionally `php-tideways` is pinned to
`php8.1-tideways`. Because sury provides other versions of PHP as well,
`update-alternatives` is used to force the 8.1 binaries to be used when no
version is specified.

Note that this has only been tested with a few extensions.

SEL-268