Update PluggableAuth to v. 7.0
Fixes #391.

This requires updating five extensions in all: PluggableAuth, LDAPAuthentication2, OpenID Connect, WSOAuth, and (because the updated version of LDAPAuthentication2 requires an updated version of it) LDAPProvider. For all five of these, the seemingly working new version can still be found in the REL1_39 branch, so there was no need to switch to master for any of them.