Version 3.0: massive overhaul, with most code moved to CanastaBase repo (#503)
* Delete _sources/canasta directory

* Delete _sources directory

* Update Dockerfile

* Update VERSION

* Delete main.cf

* Delete hadolint.yaml

* Delete images directory

* Update RELEASE_NOTES.md

* Create cirrus-search-maintenance.sh

* Create contents.yaml

* Create getSMWSettings.php

* Create smw-maintenance.sh

* Restore hadolint.yaml

* Update Dockerfile

* Re-delete hadolint.yaml

* Update docker-image.yml

* Update RELEASE_NOTES.md

* Update VERSION

* Delete .github/ISSUE_TEMPLATE directory

The main functionality (and main source of bugs, presumably) is now in CanastaBase - while the set of extensions and skins is mostly determined by what's in "Recommended revisions". So there's no real need for any of these issue templates here - and they may just confuse users.

* Update release date to today

* "Recommended Revisions" is now capitalized

It's no longer "Recommended revisions". The old URL still works, but it's better to be fully correct.

* Update RELEASE_NOTES.md

* Update RELEASE_NOTES.md

* Update Recommended Revisions URL (newer Cargo version)